refactor: refactor the frontend structure to match the nextjs best practice
