+++ frontend/app/(routes)/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import "@/components/global.css";
+import "@/_components/global.css";
 
 const inter = Inter({ subsets: ["latin"] });
