+++ frontend/app/_components/global.css
